Pulse thru using PulsePredictor class
Finished setting up the PulsePredictor class. Managed to get it passing pulses straight through to the output.

However, I only get proper triggering if I add a >=1ms delay in the loop. I suspect this is because pulses are otherwise too short to trigger my other modules. Next on the to-do list is to make an object to control pulse length.