Length pot control, jitter smoother
Wired up the length pot to control changes in cycle lengh.

There werre problems with jitter in the knob voltage - I've tried a simpe fix, but it's bugged. I think because if it throws away a new datapoint on the nth iteration, the next iteration is looking back to that datapoint which should have been thrown away.