Shuffle Control - Div/Mult mode
In Div/Mult mode, turning the encoder with the encoder switch pushed down now delays output pulses by 0 to 1 beats.

Brightness of the shuffle led indicates the extent of shuffle, but the brightness response doesn't seem very linear to me, so it's not exceptionally useful.

To do: implement CV control over enocder functions.