CV control
Fleshed out the functions which read the control voltage, and attenuvert it based on the setting of the attenuverter knob.

 There are some (hardware?) issues - e.g then the jack isn't plugged, there's quite a lot of noise on the input.

I've tidied up the way that the (length and divisions) knobs are read, to make it simpler to add the attenuated control voltage value to the knob reading directly, before scaling to the desired paramater ranges. I've maked the place in the code where that now needs doing.

Haven't started implemetation of CV control over the encoder functions (shift/shuffle).