Restore Dec31st commit
REstored the Dec 31st commit line-by line, with a few ommisions, tidying a few things up on the way.
At first, div/mult code worked ok but not CV control. In trying to fix that, I think I re-introduced the issue: it definitely seems to be related to the jitter smoother.

Jitter smoother seems to be reading the wrong input - not picking up the signal from the knob.
Removed the jitter smoothing for now.