bug hunting
I've still failed to find the source of the bug: output pulses (clock divider mode) stop while the CV attenuverter or CV in signal is rapidly altered.

I think the right PositionInOutputCycle is set by DividerMultiplier::CalculateOutputTimesMidCycle, but when I look in DividerMultiplier::ShouldWeOutputMainPulse, it often seems wrong.

One thing to consider: is it faulty that we only loop the Output cycle back to the beginning when the Input cycle resets?