Simultaneous Divider + euclidean modes
Now both the divider/multiplier and euclidean modes both run in paralellel always. The mode switch just determines whether the outputs of each make it to the the output jack. The middle setting of the mode switch outputs both, and I think this generates musically useful output.

I had to rewrite the way that the euclidean code handles shift, so that the phase of the cycles in the two different modes (as indicated by the "cycle" out) doesn't drift.

There might still be a bug somewhere which allows the phases to shift still, but I haven't managed to reproduce it, and I fixed a bug which I think might have caused the problem.

Some issues to think about:
- in the middle mode, with gate output, for most of the possible paramater space the accent is just always engaged - could it / should it be calculated differenetly?
- the new middle mode offers an alternative method of making much of the combined rhyhtms which were previously accessible only from the "thru" output. Does this reduce the justification for having the latter output? Maybe instead of a mode switch, the thru output should be replaced with a permanent gate output
- What to do in euclidean mode when divisions > length?